+++ AzStackHci.ManageUpdates/Tests/AzStackHci.ManageUpdates.Tests.ps1
@@ -34,8 +34,8 @@ Describe 'Module: AzStackHci.ManageUpdates' {
             $script:ModuleInfo | Should -Not -BeNullOrEmpty
         }
 
-        It 'Should have version 0.4.1' {
-            $script:ModuleInfo.Version | Should -Be '0.4.1'
+        It 'Should have version 0.4.2' {
+            $script:ModuleInfo.Version | Should -Be '0.4.2'
         }
 
         It 'Should export exactly 9 functions' {